+++ src/shaders/core/materials/principled.glsl
@@ -24,6 +24,17 @@ BSDFEval evalPrincipledBSDF(in Material mat, in Hit hit, in vec3 wo, in vec3 wi)
 }
 
 BSDFSample samplePrincipledBSDF(in Material mat, in Hit hit, in vec3 wo, inout uint seed) {
+    if (mat.alpha < rand(seed)) {
+        BSDFSample bsdf;
+        bsdf.wi      = -wo;
+        bsdf.pdf     = 1.0;
+        bsdf.weight  = vec3(1.0);
+        bsdf.isDelta = true;
+        bsdf.medium.isDielectric = false;
+        bsdf.medium.isVolume = false;
+        return bsdf;
+    }
+
     float pMetal = mat.metalness;
     float pTrans = (1.0 - pMetal) * mat.transmission;
     float pGlossy = 1.0 - pMetal - pTrans;
@@ -32,8 +43,14 @@ BSDFSample samplePrincipledBSDF(in Material mat, in Hit hit, in vec3 wo, inout u
     BSDFSample bsdf;
     if (r < pMetal) {
         bsdf = sampleGgxMetalBSDF(mat, hit, wo, seed);
+        bsdf.medium.isDielectric = false;
+        bsdf.medium.isVolume     = false;
     } else if (r < pMetal + pTrans) {
         bsdf = sampleDielectricBSDF(mat, hit, wo, seed);
+        bsdf.medium.absorption    = mat.albedo;
+        bsdf.medium.density       = mat.density;
+        bsdf.medium.scatterAlbedo = 0.0;
+        bsdf.medium.anisotropic   = mat.anisotropic;
     } else {
         bsdf = sampleGgxGlossyBSDF(mat, hit, wo, seed);
     }
@@ -45,10 +62,6 @@ BSDFSample samplePrincipledBSDF(in Material mat, in Hit hit, in vec3 wo, inout u
         bsdf.weight = eval.f * cosB / max(eval.pdf, EPS);
     }
 
-    bsdf.medium.isDielectric = (dot(bsdf.wi, hit.normal) < 0.0) == hit.frontFace && mat.transmission > 0.0;
-    bsdf.medium.isVolume = false;
-    bsdf.medium.absorption = mat.albedo;
-    bsdf.medium.density = mat.density;
     return bsdf;
 }